continued working on test code; found test case producing bad result
I added a small amount of code to test.guile to make it actually
make and run the executables. While making sure that this code worked,
I ran across a test case that produces bad output on the calculated
version (the case is 12 elements and 40 rotations). I've also seen
malloc fail an assertion in the brute force a number of times. I'll
worry about that problem and making the test more automated if this
test case doesn't make me pass.