Generated correct answer, eliminated main.h
Most of the files marked as "modified" were not; I suspect this is
due to moving the files off of my ext4 USB and onto my windows
installation (linux subsystem is great!).

As for the work that actually was done: I used the brute algorithm
to create the correct answer for the data that breaks the calculate
algorithm. It appears that my problem has to do with not rotating
all of the elements.

Additionally, I realized that having a 'main.h' file is smelly, and
upon further inspection it makes more sense for the sole function
declaration to be placed inside of 'shared.h' instead. Not a huge
change, but it reduces the number of files to worry about.