Made algorithm work, but runs too slowly
After fixing all of the syntax errors, it appears that the algorithm
works as it passes all of the tests that it doesn't time out on. That
being said, it only passes 3 tests so I could easily be missing
something. Either way, my next step is to lower this algorithm to c and
then begin optimizing.