Added source file as dependency; updated algorithm
The source file was not listed as a dependency of the out file,
preventing rebuilds when the source was changed; this has been fixed

The algorithm now works by going through the array only once and
tallying the number of times that a number ocurrs; the tally array is
iterated over many times. This times out 5 of the tests and for unknown
reasons produces the wrong result on all but the first (test #0) test.
I don't think this one has potential, but I want to fix it because I
might be misunderstanding the problem (which would prevent me from
seeing a more efficient solution).