updated implementation; designed interfaces
The implementation now uses indexes as pointers instead of calculations
to determine where parents and children are; this was to avoid having to
copy values around the tree. I also created defined interfaces to the
heap and heap nodes to make these sort of structural changes easier. I
still need to finish the insert algorithm (replace_node) and implement
the delete algorithm (which I'm pretty sure is pretty straight forward).